requirements.txt
openai>=1.0.0
streamlit>=1.31.0
python-dotenv>=1.0.0
requests>=2.31.0
ffmpeg-python>=0.2.0
markdown>=3.5.0
Pillow>=10.0.0
